Addresses Issue #1, as well as adding basic end-to-end tests for the converter script.
This fixes the typo that caused issue #1, and starts adding
end-to-end tests for the scripts. It does these end-to-end tests
within the pytest [http://doc.pytest.org/] framework by setting
up test cases that make subprocess calls to the converter script
and monitor the exit code as well as the resulting output
file. The output file is hashed, and compared to a master
list (kept in tests/tests_master_hashfile.txt) to ensure that the
output has not changed.

Currently this only runs three very basic conversion tests, as
well as a single test of the modify flag to the converter, on the
tests/samples/flat_sheet.json file, which includes both inserts
and deletes. This latter test would catch Issue #1 and produce an
error if it should happen again.